backend/requirements.txt
flask
psycopg2-binary
python-dotenv

gdown
sqlalchemy
sqlalchemy-utils
psycopg2-binary
pandas